Refactor to share code.
