+++ DevSkim-VSCode-Plugin/.gitignore
@@ -1,4 +1,5 @@
 .vscode-test
 devskimBinaries
-**/node_modules/**
-**/out/**
+**/node_modules
+**/out
+**/*.tsbuildinfo
