--- DevSkim-VSCode-Plugin/client/src/codeFixMapping.ts
@@ -1,14 +1,0 @@
-import * as vscode from 'vscode';
-
-export class CodeFixMapping
-{
-	diagnostic: vscode.Diagnostic;
-	replacement: string;
-	fileName: string;
-	constructor(diagnostic: vscode.Diagnostic, replacement: string, fileName: string)
-	{
-		this.diagnostic = diagnostic;
-		this.replacement = replacement;
-		this.fileName = fileName;
-	}
-}
