+++ DevSkim-VSCode-Plugin/common/languagesAccess.ts
