+++ DevSkim-VSCode-Plugin/common/selectors.ts
