--- DevSkim-VSCode-Plugin/server/src/codeFixMapping.ts
@@ -1,15 +1,0 @@
-import {
-	Diagnostic
-} from 'vscode-languageserver/node';
-export class CodeFixMapping
-{
-	diagnostic: Diagnostic;
-	replacement: string;
-	fileName: string;
-	constructor(diagnostic: Diagnostic, replacement: string, fileName: string)
-	{
-		this.diagnostic = diagnostic;
-		this.replacement = replacement;
-		this.fileName = fileName;
-	}
-}
