--- DevSkim-VSCode-Plugin/server/src/devskimSettings.ts
@@ -1,25 +1,0 @@
-/* eslint-disable @typescript-eslint/no-inferrable-types */
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-export interface DevSkimSettings {
-    enableManualReviewRules: boolean;
-    enableBestPracticeRules: boolean;
-    suppressionDurationInDays: number;
-    manualReviewerName: string;
-    ignoreFiles: string[];
-    ignoreRulesList: string;
-    guidanceBaseURL: string;
-    removeFindingsOnClose: boolean;
-}
-
-export class DevSkimSettingsObject implements DevSkimSettings {
-    enableBestPracticeRules: boolean = false;
-    enableManualReviewRules: boolean = false;
-    guidanceBaseURL: string = "https://github.com/Microsoft/DevSkim/Guidance";
-    ignoreFiles: string[] = [];
-    ignoreRulesList: string = "";
-    manualReviewerName: string = "";
-    removeFindingsOnClose: boolean = false;
-    suppressionDurationInDays: number = 30;
-}
